pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ticktick-mcp"
version = "0.1.0"
description = "MCP (Model Context Protocol) server for TickTick task management"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Ainur Nigmatullin" }
]
keywords = [
    "mcp",
    "model-context-protocol",
    "ticktick",
    "tasks",
    "productivity",
    "ai",
    "claude",
    "llm",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/ainigmatullin/ticktick-mcp"
Repository = "https://github.com/ainigmatullin/ticktick-mcp"
Issues = "https://github.com/ainigmatullin/ticktick-mcp/issues"

[project.scripts]
ticktick-mcp = "ticktick_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ticktick_mcp"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
